+++ changelog.md
@@ -22,7 +22,7 @@
 * **[AirWing]** Use aircraft display names for easier differentiation between modules
 
 ## Fixes
-* **[Flight Plans]** Fixed a bug when a package was created with only escort flights
+* **[Flight Plans]** Fixed bugs wrt planning escort flights
 * **[Flight Plans]** Added AntiShipStrike as a fallback task for OCA/Aircraft to fix a bug where the S-3B could not do OCA/Aircraft
 * **[Squadrons]** Fixed a bug where loading an air wing config would not properly load all squadrons
 * **[Flight Plans]** Fixed a bug where SEAD flights would fire one ARM and RTB